+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/session/SessionComponent.kt
@@ -19,11 +19,11 @@ package im.vector.matrix.android.internal.session
 import dagger.BindsInstance
 import dagger.Component
 import im.vector.matrix.android.api.auth.data.SessionParams
-import im.vector.matrix.android.api.session.InitialSyncProgressService
 import im.vector.matrix.android.api.session.Session
 import im.vector.matrix.android.internal.crypto.CryptoModule
 import im.vector.matrix.android.internal.di.MatrixComponent
 import im.vector.matrix.android.internal.network.NetworkConnectivityChecker
+import im.vector.matrix.android.internal.session.user.accountdata.AccountDataModule
 import im.vector.matrix.android.internal.session.cache.CacheModule
 import im.vector.matrix.android.internal.session.content.ContentModule
 import im.vector.matrix.android.internal.session.content.UploadContentWorker
@@ -46,20 +46,21 @@ import im.vector.matrix.android.internal.session.user.UserModule
 import im.vector.matrix.android.internal.task.TaskExecutor
 
 @Component(dependencies = [MatrixComponent::class],
-        modules = [
-            SessionModule::class,
-            RoomModule::class,
-            SyncModule::class,
-            SignOutModule::class,
-            GroupModule::class,
-            UserModule::class,
-            FilterModule::class,
-            GroupModule::class,
-            ContentModule::class,
-            CacheModule::class,
-            CryptoModule::class,
-            PushersModule::class
-        ]
+           modules = [
+               SessionModule::class,
+               RoomModule::class,
+               SyncModule::class,
+               SignOutModule::class,
+               GroupModule::class,
+               UserModule::class,
+               FilterModule::class,
+               GroupModule::class,
+               ContentModule::class,
+               CacheModule::class,
+               CryptoModule::class,
+               PushersModule::class,
+               AccountDataModule::class
+           ]
 )
 @SessionScope
 internal interface SessionComponent {