+++ matrix-sdk-android/src/main/java/im/vector/matrix/android/internal/util/StringUtils.kt
@@ -16,6 +16,7 @@
 
 package im.vector.matrix.android.internal.util
 
+import im.vector.matrix.android.api.MatrixPatterns
 import timber.log.Timber
 
 /**
@@ -48,4 +49,11 @@ fun convertFromUTF8(s: String): String? {
         Timber.e(e, "## convertFromUTF8()  failed")
         null
     }
+}
+
+fun String?.firstLetterOfDisplayName(): String {
+    if (this.isNullOrEmpty()) return ""
+    val isUserId = MatrixPatterns.isUserId(this)
+    val firstLetterIndex = if (isUserId) 1 else 0
+    return this[firstLetterIndex].toString().toUpperCase()
 }